Code now works with spelling mistakes as well

+++ src/main/java/org/group1/response/distances/Demerau.java
@@ -3,7 +3,7 @@ package org.group1.response.distances;
 public class Demerau implements iDistance {
 
     @Override
-    public float calculate(String a, String b) {
+    public double calculate(String a, String b) {
         return 0.0090980f;
     }
 }
