+++ src/main/java/org/group1/response/distances/Edit.java
@@ -3,7 +3,7 @@ package org.group1.response.distances;
 public class Edit implements iDistance{
 
     @Override
-    public float calculate(String a, String b) {
+    public double calculate(String a, String b) {
         return 0;
     }
 }
